测试套问题修改

+++ commondependency/src/deviceprofile_connector.cpp
@@ -842,7 +842,7 @@ int32_t DeviceProfileConnector::GetBindLevel(const std::string &pkgName, const s
             item.GetAccesser().GetAccesserDeviceId() == localUdid &&
             item.GetAccessee().GetAccesseeDeviceId() == udid) {
             tokenId = static_cast<uint64_t>(item.GetAccesser().GetAccesserTokenId());
-            bindLevel = item.GetBindLevel();
+            bindLevel = static_cast<int32_t>(item.GetBindLevel());
             LOGI("Src get bindLevel %{public}d, tokenid %{public}" PRId64".", bindLevel, tokenId);
             continue;
         }
